refactor: extract shared types, utils, hooks, and remove dead code
- Extract CarouselPostRequired type to lib/types.ts
- Extract formatDate, extractImagesFromHtml, calculateReadingTime to lib/utils.ts
- Extract useCarouselScroll hook (RAF, drag, inertia, keyboard)
- Extract CarouselInfoPanel component (right panel with AnimatePresence)
- PostCarousel2D reduced from 369 to 118 lines
- Remove unused files: MainPage, PostContent, LiquidGlassTest, CursorWrapper
