+++ app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import "./globals.css";
 import { NavBar } from "@/components/NavBar";
-// import { CursorWrapper } from "@/components/CursorWrapper";
 
 export const metadata: Metadata = {
   title: "Minkyo Jung",
@@ -16,10 +15,8 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <body className="antialiased">
-        {/* <CursorWrapper> */}
-          <NavBar />
-          {children}
-        {/* </CursorWrapper> */}
+        <NavBar />
+        {children}
       </body>
     </html>
   );
